l10n_fr_das2: move declaration threshold to pyfrdas2
Move the partner declaration threashold to the pyfrdas2 lib. This
configuration parameter should never be changed by the use anyway. With
the encryption that change every year, we will release every year a new
lib, so we can change the result of get_partner_declaration_threshold()
if the tax administration changes the amount.